Moved main functions out to the ML/Base.py file
