+++ teaspoon/ML/tents.py
@@ -36,171 +36,6 @@ import itertools
 
 
 
-## 	A class for storing all of the chosen parameters for a data frame and computation of ML.
-#
-class ParameterBucket(object):
-	def __init__(self, description = '',
-					d=10,
-					delta = 1,
-					epsilon = 0,
-					maxPower = 1,
-					clfClass = RidgeClassifierCV,
-					seed = None,
-					test_size = .33,
-                 	feature_function=None,
-                 	boundingBoxMatrix = None):
-		"""!@brief Creates a new ParameterBucket object.
-
-	    This object is being used to keep track of all the parameters needed
-	    for the tents ML featurization.
-
-	    Parameters that are included in the ParameterBucket initially:
-
-	    @param description A description, has no effect on code. This can be set on initialization.
-	    @param d, delta, epsilon
-	    	The bounding box for the persistence diagram in the (birth, lifetime) coordinates is [0,d * delta] x [epsilon, d* delta + epsilon].  In the usual coordinates, this creates a parallelogram.
-	    @param maxPower
-	    	The maximum degree used for the monomial combinations of the tent functions.  Testing suggests we usually want this to be 1.  Increasing causes large increase in number of features.
-	    @param clfClass
-	    	The choice of tool used for classification or regression, passed as the function.  This code has been tested using `sklearn` functions `RidgeClassiferCV` for classification and `RidgeCV` for regression.
-	    @param seed
-	    	The seed for the pseudo-random number generator.  Pass None if you don't want it fixed; otherwise, pass an integer.
-	    @param test_size
-	    	A number in \f$[0,1]\f$.  Gives the percentage of data points to be reserved for the testing set if this is being used for a train/test split experiment.  Otherwise, ignored.
-        @param feature_function
-	    	The basis function you want to use for interpolation. Default is tent()
-	    @param boundingBoxMatrix
-	    	Not yet implemented.  See self.findBoundingBox()
-
-
-	    """
-		self.description = description
-		self.d = d
-		self.delta = delta
-		self.epsilon = epsilon
-
-		self.maxPower = maxPower
-		self.clfClass = clfClass
-
-		self.seed = seed
-
-		self.test_size = test_size
-
-		# @todo The following settings don't appear to be ever used.  Commented in case that's not correct, but should be removed eventually.
-		# self.minBirth = None
-		# self.maxBirth = None
-		# self.minPers = None
-		# self.maxPers = None
-		# self.remove0cols = False
-		if feature_function == None:
-			self.feature_function = fF.tent
-		else:
-			self.feature_function = feature_function
-
-
-	def __str__(self):
-		"""!
-		@brief Nicely prints all currently set values in the ParameterBucket.
-		"""
-		attrs = vars(self)
-		output = ''
-		output += 'Variables in parameter bucket\n'
-		output += '---\n'
-		for key in attrs.keys():
-			output += str(key) + ' : '
-			output += str(attrs[key])+ '\n'
-		output += '---\n'
-		return output
-
-	def encloseDgms(self, DgmSeries):
-		'''!
-		@brief Tests to see if the parameters enclose the persistence diagrams in the DgmSeries
-
-		@returns boolean
-		'''
-
-		# Height of parallelogram; equivalently maximum lifetime enclosed
-		height = self.d * self.delta + self.epsilon
-		width = self.d * self.delta
-
-		minBirth = pP.minBirthSeries(DgmSeries)
-		if minBirth <0:
-			print("This code assumes positive birth times.")
-			return False
-
-		maxBirth = pP.maxBirthSeries(DgmSeries)
-		if maxBirth > width:
-			print('There are birth times outside the bounding box.')
-			return False
-
-		minPers = pP.minPersistenceSeries(DgmSeries)
-		if minPers < self.epsilon:
-			print('There are points below the epsilon shift.')
-			return False
-
-		maxPers = pP.maxPersistenceSeries(DgmSeries)
-		if maxPers > height:
-			print('There are points above the box.')
-			return False
-
-		return True
-
-	def chooseDeltaEpsWithPadding(self, DgmsSeries, pad = 0):
-		'''
-
-		DgmsSeries is pd.series
-		d is number of grid elements in either direction
-		pad is the additional padding outside of the points in the diagrams
-
-		Sets the needed delta and epsilon
-
-
-		'''
-
-		topPers = pP.maxPersistenceSeries(DgmsSeries)
-		bottomPers = pP.minPersistenceSeries(DgmsSeries)
-		topBirth = max(DgmsSeries.apply(pP.maxBirth))
-
-		height = max(topPers,topBirth)
-
-		bottomBirth = min(DgmsSeries.apply(pP.minBirth))
-		if bottomBirth < 0:
-			print('This code assumes that birth time is always positive\nbut you have negative birth times....')
-			print('Your minimum birth time was', bottomBirth)
-
-		epsilon = bottomPers/2
-
-		delta = (height + pad - epsilon) / self.d
-
-
-		self.delta = delta
-		self.epsilon = epsilon
-
-	def findBoundingBox(self,DgmsSeries,pad = 0):
-		'''
-		DgmsSeries is of type pd.series
-		pad is the additional padding outside of the points in the diagrams
-
-
-		Sets a bounding box in the birth-lifetime plane
-		to use for creating support of function collection.
-
-		Result is `self.boundingBox` is a dictionary with
-		two keys, 'birthAxis' and 'lifetimeAxis', each outputing
-		a tuple of length 2.
-
-
-		'''
-		topPers = pP.maxPersistenceSeries(DgmsSeries)
-		bottomPers = pP.minPersistenceSeries(DgmsSeries)
-		topBirth = max(DgmsSeries.apply(pP.maxBirth))
-		bottomBirth = min(DgmsSeries.apply(pP.minBirth))
-
-
-		self.boundingBox = {}
-		self.boundingBox['birthAxis'] = (bottomBirth - pad, topBirth + pad)
-		self.boundingBox['lifetimeAxis'] = (bottomPers/2, topPers + pad)
-
 
 
 
@@ -299,7 +134,7 @@ def tent(Dgm, params, type = 'BirthDeath'):
 def build_G(DgmSeries, params):
     applyTents = lambda x: params.feature_function(x,params = params)
     G = np.array(list(DgmSeries.apply(applyTents )))
-    
+
     # Include powers if necessary
     if params.maxPower>1:
         poly = PolynomialFeatures(params.maxPower)
@@ -361,30 +196,30 @@ def TentML(DgmsDF,
         # print('params.maxPower = ', params.maxPower)
         print('Exiting....')
         return
-    
+
     clf = params.clfClass()
-    
+
     if verbose:
         print('Training estimator.')
-        
+
 #    startTime = time.time()
-    
+
     #check to see if only one column label was passed. If so, turn it into a list.
     if type(dgm_col) == str:
         dgm_col = [dgm_col]
-        
+
     if verbose:
         print('Making G...')
-        
+
     listOfG = []
     for dgmColLabel in dgm_col:
         G = build_G(DgmsDF[dgmColLabel],params)
         listOfG.append(G)
-    
+
     G = np.concatenate(listOfG,axis = 1)
 
     numFeatures = np.shape(G)[1]
-     
+
     # Normalize G
     if normalize:
         G = scale(G)
@@ -392,130 +227,20 @@ def TentML(DgmsDF,
 
     if verbose:
         print('Number of features used is', numFeatures,'...')
-        
+
     clf.fit(G,list(DgmsDF[labels_col]))
-    
+
     if verbose:
         print('Checking score on training set...')
-        
+
     score = clf.score(G,list(DgmsDF[labels_col]))
-    
+
     if verbose:
         print('Score on training set: ' + str(score) + '.\n')
-        
+
     clf.delta = params.delta
     clf.epsilon = params.epsilon
     clf.trainingScore = score
     clf.d = params.d
-    
+
     return clf
-
-
-
-
-## Main testing function for classification or regression methods.
-# Does train/test split, creates classifier, and returns score on test.
-#
-# 	@param DgmsDF
-# 		A pandas data frame containing, at least, a column of
-# 		diagrams and a column of labels
-# 	@param labels_col
-# 		A string.  The label for the column in DgmsDF containing the training labels.
-# 	@param dgm_col
-# 		A string or list of strings giving the label for the column containing the diagrams.
-# 	@param params
-# 		A class of type ParameterBucket
-# 		Should store:
-# 			- **d**:
-# 				An integer, the number of elements for griding up
-# 				the x and y axis of the diagram.  Will result in
-# 				d*(d+1) tent functions
-# 			- **delta**, **epsilon**:
-# 				Controls location and width of mesh elements for x and y axis of the
-# 				diagram.
-# 			- **clfClass**:
-# 				The class which will be used for classification.  Currently tested
-#				using `sklearn.RidgeClassifierCV` and `sklearn.RidgeCV`.
-#			- **seed**:
-#				None if we don't want to mess with the seed for the train_test_split function. Else, pass integer.
-#			- **test_split**:
-#				The percentage of the data to be reserved for the test part of the train/test split.
-#
-# 	@return
-#		Returned as a dictionary of entries:
-# 		- **score**
-# 			The percent correct when predicting on the test set.
-# 		- **DgmsDF**
-# 			The original data frame passed back with a column labeled
-# 			'Prediction' added with the predictions gotten for the
-# 			test set. Data points in the training set will have an
-# 			entry of NaN
-# 		- **clf**
-# 			The classifier object.
-#
-def getPercentScore(DgmsDF,
-					labels_col = 'trainingLabel',
-					dgm_col = 'Dgm1',
-					params = ParameterBucket(),
-					normalize = False,
-					verbose = True
-					):
-
-	print('---')
-	print('Beginning experiment.')
-	if verbose:
-		print(params)
-
-	#check to see if only one column label was passed. If so, turn it into a list.
-	if type(dgm_col) == str:
-		dgm_col = [dgm_col]
-
-	# Run actual train/test experiment using sklearn
-	D_train, D_test, L_train,L_test = train_test_split(DgmsDF,
-													DgmsDF[labels_col],
-													test_size=params.test_size,
-													random_state = params.seed
-													)
-
-	#--------Training------------#
-	if verbose:
-		print('Using ' + str(len(L_train)) + '/' + str(len(DgmsDF)) + ' to train...')
-	clf = TentML(D_train,
-					labels_col = labels_col,
-					dgm_col = dgm_col,
-					params = params,
-					normalize = normalize,
-					verbose = verbose)
-
-	#--------Testing-------------#
-	if verbose:
-		print('Using ' + str(len(L_test)) + '/' + str(len(DgmsDF)) + ' to test...')
-	listOfG = []
-	for dgmColLabel in dgm_col:
-		G = build_G(D_test[dgmColLabel],params)
-		listOfG.append(G)
-
-	G = np.concatenate(listOfG,axis = 1)
-
-	# Normalize G
-	if normalize:
-		G = scale(G)
-
-
-	# Compute predictions and add to DgmsDF data frame
-	L_predict = pd.Series(clf.predict(G),index = L_test.index)
-	DgmsDF['Prediction'] = L_predict
-
-	# Compute score
-	score = clf.score(G,list(L_test))
-	if verbose:
-		print('Score on testing set: ' + str(score) +"...\n")
-
-	print('Finished with train/test experiment.')
-
-	output = {}
-	output['score'] = score
-	output['DgmsDF'] = DgmsDF
-	output['clf'] = clf
-
-	return output
